Clarify wording around CLA not being received
The need to wait at least one *business* day is explained in the
devguide, but it's not too prominent there.
The Knights' current wording makes it sound like the ball is in
the contributor's court, but it's not. If someone signs the CLA
on a Saturday, they actually need to wait for the whole weekend
before the contribution can be reviewed.

- Change "you have not signed the CLA" to "we have not received your CLA".

- Add a note on waiting for a businnes day directly to the commit template.