Mention which users have not signed the CLA
Update `check` and `update` to include in the message a list of users
that either didn't sign the CLA or haven't been found.

The implementation follows [@brettcannon's
suggestion](https://github.com/python/the-knights-who-say-ni/issues/128#issuecomment-337313178)
to return a `Mapping[Status, AbstractSet[str]]` from `check` that is a
map of "problematic" statuses and their associated list of usernames.