Make usernames for NO_USERNAME_BODY bpo-friendly
When the bot posts a message for a PR author, there is a risk that
the author will copy a username inside to their bpo profile settings
as is, with initial `@`. As a result, the bot will not recognize/report
an update hence creating confusion.

Since presence of `@` adds no value (the author gets a notification
anyway), this commit safely turns this:

> We couldn't find a bugs.python.org (b.p.o) [...]:
>
> @username

into this:

> We couldn't find a bugs.python.org (b.p.o) [...]:
>
> **username**