Add useful links to the Knight's message
Those who are greeted by the Knight on GitHub are typically people contributing to CPython for the first time. Often, not having signed the CLA is symptomatic of the new contributor not having read any of the Developer's Guide, which may mean that there are other problems with their PR that make it difficult for it to be merged.

This PR proposes adding some helpful links to the Knight's message, pointing new contributors towards the Developer's Guide.